@@ -27,12 +27,11 @@ This module contains different mixin classes for internal use within the Amazon 
 
 from __future__ import annotations
 
-from functools import cached_property
+from functools import cache, cached_property
 from typing import Any, Generic, NamedTuple, TypeVar
 
 from typing_extensions import final
 
-from airflow.compat.functools import cache
 from airflow.providers.amazon.aws.hooks.base_aws import AwsGenericHook
 
 AwsHookType = TypeVar("AwsHookType", bound=AwsGenericHook)
