@@ -19,6 +19,7 @@ from __future__ import annotations
 import logging
 import secrets
 import string
+from functools import cache
 from typing import TYPE_CHECKING
 
 import pendulum
@@ -26,7 +27,6 @@ from deprecated import deprecated
 from kubernetes.client.rest import ApiException
 from slugify import slugify
 
-from airflow.compat.functools import cache
 from airflow.configuration import conf
 from airflow.exceptions import AirflowProviderDeprecationWarning
 
