@@ -25,9 +25,10 @@ import shutil
 import subprocess
 import sys
 import tempfile
+from collections.abc import Iterable
 from io import StringIO
 from pathlib import Path
-from typing import Any, Iterable, NamedTuple
+from typing import Any, NamedTuple
 
 import click
 
