@@ -17,14 +17,14 @@
 from __future__ import annotations
 
 import os
-from functools import lru_cache
+from functools import cache
 
 from black import Mode, TargetVersion, format_str, parse_pyproject_toml
 
 from airflow_breeze.utils.path_utils import AIRFLOW_SOURCES_ROOT
 
 
-@lru_cache(maxsize=None)
+@cache
 def _black_mode() -> Mode:
     config = parse_pyproject_toml(os.path.join(AIRFLOW_SOURCES_ROOT, "pyproject.toml"))
     target_versions = {TargetVersion[val.upper()] for val in config.get("target_version", ())}
