@@ -23,7 +23,7 @@ from __future__ import annotations
 
 import os
 from enum import Enum
-from functools import lru_cache
+from functools import cache
 from typing import NamedTuple, TextIO
 
 from rich.console import Console
@@ -90,7 +90,7 @@ class Output(NamedTuple):
         return self.title.replace("[", "\\[")
 
 
-@lru_cache(maxsize=None)
+@cache
 def get_console(output: Output | None = None) -> Console:
     return Console(
         force_terminal=True,
@@ -102,7 +102,7 @@ def get_console(output: Output | None = None) -> Console:
     )
 
 
-@lru_cache(maxsize=None)
+@cache
 def get_stderr_console(output: Output | None = None) -> Console:
     return Console(
         force_terminal=True,
