@@ -18,7 +18,7 @@ from __future__ import annotations
 
 import json
 import re
-from functools import lru_cache
+from functools import cache
 from typing import Any
 
 import pytest
@@ -56,8 +56,7 @@ def escape_ansi_colors(line):
     return ANSI_COLORS_MATCHER.sub("", line)
 
 
-# Can be replaced with cache when we move to Python 3.9 (when 3.8 is EOL)
-@lru_cache(maxsize=None)
+@cache
 def get_rich_console() -> Console:
     return Console(color_system="truecolor", force_terminal=True)
 
@@ -111,10 +110,10 @@ def assert_outputs_are_printed(expected_outputs: dict[str, str], stderr: str):
                 ("INTHEWILD.md",),
                 {
                     "affected-providers-list-as-string": None,
-                    "all-python-versions": "['3.8']",
-                    "all-python-versions-list-as-string": "3.8",
-                    "python-versions": "['3.8']",
-                    "python-versions-list-as-string": "3.8",
+                    "all-python-versions": "['3.9']",
+                    "all-python-versions-list-as-string": "3.9",
+                    "python-versions": "['3.9']",
+                    "python-versions-list-as-string": "3.9",
                     "ci-image-build": "false",
                     "needs-helm-tests": "false",
                     "run-tests": "false",
@@ -137,10 +136,10 @@ def assert_outputs_are_printed(expected_outputs: dict[str, str], stderr: str):
                 ("airflow/api/file.py",),
                 {
                     "affected-providers-list-as-string": "common.compat fab",
-                    "all-python-versions": "['3.8']",
-                    "all-python-versions-list-as-string": "3.8",
-                    "python-versions": "['3.8']",
-                    "python-versions-list-as-string": "3.8",
+                    "all-python-versions": "['3.9']",
+                    "all-python-versions-list-as-string": "3.9",
+                    "python-versions": "['3.9']",
+                    "python-versions-list-as-string": "3.9",
                     "ci-image-build": "true",
                     "prod-image-build": "false",
                     "needs-helm-tests": "false",
@@ -163,10 +162,10 @@ def assert_outputs_are_printed(expected_outputs: dict[str, str], stderr: str):
             pytest.param(
                 ("airflow/api_internal/file.py",),
                 {
-                    "all-python-versions": "['3.8']",
-                    "all-python-versions-list-as-string": "3.8",
-                    "python-versions": "['3.8']",
-                    "python-versions-list-as-string": "3.8",
+                    "all-python-versions": "['3.9']",
+                    "all-python-versions-list-as-string": "3.9",
+                    "python-versions": "['3.9']",
+                    "python-versions-list-as-string": "3.9",
                     "ci-image-build": "true",
                     "prod-image-build": "false",
                     "needs-helm-tests": "false",
@@ -188,10 +187,10 @@ def assert_outputs_are_printed(expected_outputs: dict[str, str], stderr: str):
             pytest.param(
                 ("airflow/api_fastapi/file.py",),
                 {
-                    "all-python-versions": "['3.8']",
-                    "all-python-versions-list-as-string": "3.8",
-                    "python-versions": "['3.8']",
-                    "python-versions-list-as-string": "3.8",
+                    "all-python-versions": "['3.9']",
+                    "all-python-versions-list-as-string": "3.9",
+                    "python-versions": "['3.9']",
+                    "python-versions-list-as-string": "3.9",
                     "ci-image-build": "true",
                     "prod-image-build": "false",
                     "needs-helm-tests": "false",
@@ -213,10 +212,10 @@ def assert_outputs_are_printed(expected_outputs: dict[str, str], stderr: str):
             pytest.param(
                 ("tests/api/file.py",),
                 {
-                    "all-python-versions": "['3.8']",
-                    "all-python-versions-list-as-string": "3.8",
-                    "python-versions": "['3.8']",
-                    "python-versions-list-as-string": "3.8",
+                    "all-python-versions": "['3.9']",
+                    "all-python-versions-list-as-string": "3.9",
+                    "python-versions": "['3.9']",
+                    "python-versions-list-as-string": "3.9",
                     "ci-image-build": "true",
                     "prod-image-build": "false",
                     "needs-helm-tests": "false",
@@ -239,10 +238,10 @@ def assert_outputs_are_printed(expected_outputs: dict[str, str], stderr: str):
                 ("airflow/operators/file.py",),
                 {
                     "affected-providers-list-as-string": None,
-                    "all-python-versions": "['3.8']",
-                    "all-python-versions-list-as-string": "3.8",
-                    "python-versions": "['3.8']",
-                    "python-versions-list-as-string": "3.8",
+                    "all-python-versions": "['3.9']",
+                    "all-python-versions-list-as-string": "3.9",
+                    "python-versions": "['3.9']",
+                    "python-versions-list-as-string": "3.9",
                     "ci-image-build": "true",
                     "prod-image-build": "false",
                     "needs-helm-tests": "false",
@@ -266,10 +265,10 @@ def assert_outputs_are_printed(expected_outputs: dict[str, str], stderr: str):
                 ("airflow/operators/python.py",),
                 {
                     "affected-providers-list-as-string": None,
-                    "all-python-versions": "['3.8']",
-                    "all-python-versions-list-as-string": "3.8",
-                    "python-versions": "['3.8']",
-                    "python-versions-list-as-string": "3.8",
+                    "all-python-versions": "['3.9']",
+                    "all-python-versions-list-as-string": "3.9",
+                    "python-versions": "['3.9']",
+                    "python-versions-list-as-string": "3.9",
                     "ci-image-build": "true",
                     "prod-image-build": "false",
                     "needs-helm-tests": "false",
@@ -295,10 +294,10 @@ def assert_outputs_are_printed(expected_outputs: dict[str, str], stderr: str):
                 ("airflow/serialization/python.py",),
                 {
                     "affected-providers-list-as-string": None,
-                    "all-python-versions": "['3.8']",
-                    "all-python-versions-list-as-string": "3.8",
-                    "python-versions": "['3.8']",
-                    "python-versions-list-as-string": "3.8",
+                    "all-python-versions": "['3.9']",
+                    "all-python-versions-list-as-string": "3.9",
+                    "python-versions": "['3.9']",
+                    "python-versions-list-as-string": "3.9",
                     "ci-image-build": "true",
                     "prod-image-build": "false",
                     "needs-helm-tests": "false",
@@ -326,10 +325,10 @@ def assert_outputs_are_printed(expected_outputs: dict[str, str], stderr: str):
                 {
                     "affected-providers-list-as-string": "amazon common.compat common.sql fab google openlineage "
                     "pgvector postgres",
-                    "all-python-versions": "['3.8']",
-                    "all-python-versions-list-as-string": "3.8",
-                    "python-versions": "['3.8']",
-                    "python-versions-list-as-string": "3.8",
+                    "all-python-versions": "['3.9']",
+                    "all-python-versions-list-as-string": "3.9",
+                    "python-versions": "['3.9']",
+                    "python-versions-list-as-string": "3.9",
                     "ci-image-build": "true",
                     "prod-image-build": "false",
                     "needs-helm-tests": "false",
@@ -357,10 +356,10 @@ def assert_outputs_are_printed(expected_outputs: dict[str, str], stderr: str):
                 ("tests/providers/apache/beam/file.py",),
                 {
                     "affected-providers-list-as-string": "apache.beam google",
-                    "all-python-versions": "['3.8']",
-                    "all-python-versions-list-as-string": "3.8",
-                    "python-versions": "['3.8']",
-                    "python-versions-list-as-string": "3.8",
+                    "all-python-versions": "['3.9']",
+                    "all-python-versions-list-as-string": "3.9",
+                    "python-versions": "['3.9']",
+                    "python-versions-list-as-string": "3.9",
                     "ci-image-build": "true",
                     "prod-image-build": "false",
                     "needs-helm-tests": "false",
@@ -385,10 +384,10 @@ def assert_outputs_are_printed(expected_outputs: dict[str, str], stderr: str):
                 ("docs/file.rst",),
                 {
                     "affected-providers-list-as-string": None,
-                    "all-python-versions": "['3.8']",
-                    "all-python-versions-list-as-string": "3.8",
-                    "python-versions": "['3.8']",
-                    "python-versions-list-as-string": "3.8",
+                    "all-python-versions": "['3.9']",
+                    "all-python-versions-list-as-string": "3.9",
+                    "python-versions": "['3.9']",
+                    "python-versions-list-as-string": "3.9",
                     "ci-image-build": "true",
                     "prod-image-build": "false",
                     "needs-helm-tests": "false",
@@ -416,10 +415,10 @@ def assert_outputs_are_printed(expected_outputs: dict[str, str], stderr: str):
                 {
                     "affected-providers-list-as-string": "amazon common.sql google openlineage "
                     "pgvector postgres",
-                    "all-python-versions": "['3.8']",
-                    "all-python-versions-list-as-string": "3.8",
-                    "python-versions": "['3.8']",
-                    "python-versions-list-as-string": "3.8",
+                    "all-python-versions": "['3.9']",
+                    "all-python-versions-list-as-string": "3.9",
+                    "python-versions": "['3.9']",
+                    "python-versions-list-as-string": "3.9",
                     "ci-image-build": "true",
                     "prod-image-build": "true",
                     "needs-helm-tests": "true",
@@ -451,10 +450,10 @@ def assert_outputs_are_printed(expected_outputs: dict[str, str], stderr: str):
                 {
                     "affected-providers-list-as-string": "amazon apache.livy "
                     "dbt.cloud dingding discord http",
-                    "all-python-versions": "['3.8']",
-                    "all-python-versions-list-as-string": "3.8",
-                    "python-versions": "['3.8']",
-                    "python-versions-list-as-string": "3.8",
+                    "all-python-versions": "['3.9']",
+                    "all-python-versions-list-as-string": "3.9",
+                    "python-versions": "['3.9']",
+                    "python-versions-list-as-string": "3.9",
                     "ci-image-build": "true",
                     "prod-image-build": "true",
                     "needs-helm-tests": "true",
@@ -487,10 +486,10 @@ def assert_outputs_are_printed(expected_outputs: dict[str, str], stderr: str):
                 ),
                 {
                     "affected-providers-list-as-string": "airbyte",
-                    "all-python-versions": "['3.8']",
-                    "all-python-versions-list-as-string": "3.8",
-                    "python-versions": "['3.8']",
-                    "python-versions-list-as-string": "3.8",
+                    "all-python-versions": "['3.9']",
+                    "all-python-versions-list-as-string": "3.9",
+                    "python-versions": "['3.9']",
+                    "python-versions-list-as-string": "3.9",
                     "ci-image-build": "true",
                     "prod-image-build": "true",
                     "needs-helm-tests": "true",
@@ -519,10 +518,10 @@ def assert_outputs_are_printed(expected_outputs: dict[str, str], stderr: str):
                 ),
                 {
                     "affected-providers-list-as-string": None,
-                    "all-python-versions": "['3.8']",
-                    "all-python-versions-list-as-string": "3.8",
-                    "python-versions": "['3.8']",
-                    "python-versions-list-as-string": "3.8",
+                    "all-python-versions": "['3.9']",
+                    "all-python-versions-list-as-string": "3.9",
+                    "python-versions": "['3.9']",
+                    "python-versions-list-as-string": "3.9",
                     "ci-image-build": "true",
                     "prod-image-build": "true",
                     "needs-helm-tests": "true",
@@ -547,10 +546,10 @@ def assert_outputs_are_printed(expected_outputs: dict[str, str], stderr: str):
                 ("generated/provider_dependencies.json",),
                 {
                     "affected-providers-list-as-string": ALL_PROVIDERS_AFFECTED,
-                    "all-python-versions": "['3.8', '3.9', '3.10', '3.11', '3.12']",
-                    "all-python-versions-list-as-string": "3.8 3.9 3.10 3.11 3.12",
-                    "python-versions": "['3.8', '3.9', '3.10', '3.11', '3.12']",
-                    "python-versions-list-as-string": "3.8 3.9 3.10 3.11 3.12",
+                    "all-python-versions": "['3.9', '3.10', '3.11', '3.12']",
+                    "all-python-versions-list-as-string": "3.9 3.10 3.11 3.12",
+                    "python-versions": "['3.9', '3.10', '3.11', '3.12']",
+                    "python-versions-list-as-string": "3.9 3.10 3.11 3.12",
                     "ci-image-build": "true",
                     "prod-image-build": "true",
                     "needs-helm-tests": "true",
@@ -574,10 +573,10 @@ def assert_outputs_are_printed(expected_outputs: dict[str, str], stderr: str):
                 ("generated/provider_dependencies.json",),
                 {
                     "affected-providers-list-as-string": ALL_PROVIDERS_AFFECTED,
-                    "all-python-versions": "['3.8', '3.9', '3.10', '3.11', '3.12']",
-                    "all-python-versions-list-as-string": "3.8 3.9 3.10 3.11 3.12",
-                    "python-versions": "['3.8', '3.9', '3.10', '3.11', '3.12']",
-                    "python-versions-list-as-string": "3.8 3.9 3.10 3.11 3.12",
+                    "all-python-versions": "['3.9', '3.10', '3.11', '3.12']",
+                    "all-python-versions-list-as-string": "3.9 3.10 3.11 3.12",
+                    "python-versions": "['3.9', '3.10', '3.11', '3.12']",
+                    "python-versions-list-as-string": "3.9 3.10 3.11 3.12",
                     "ci-image-build": "true",
                     "prod-image-build": "true",
                     "needs-helm-tests": "true",
@@ -601,10 +600,10 @@ def assert_outputs_are_printed(expected_outputs: dict[str, str], stderr: str):
                 "affected-providers-list-as-string": "amazon apache.hive cncf.kubernetes "
                 "common.compat common.sql exasol ftp google http imap microsoft.azure "
                 "mongo mysql openlineage postgres salesforce ssh teradata",
-                "all-python-versions": "['3.8']",
-                "all-python-versions-list-as-string": "3.8",
-                "python-versions": "['3.8']",
-                "python-versions-list-as-string": "3.8",
+                "all-python-versions": "['3.9']",
+                "all-python-versions-list-as-string": "3.9",
+                "python-versions": "['3.9']",
+                "python-versions-list-as-string": "3.9",
                 "ci-image-build": "true",
                 "prod-image-build": "false",
                 "needs-helm-tests": "false",
@@ -627,10 +626,10 @@ def assert_outputs_are_printed(expected_outputs: dict[str, str], stderr: str):
             ("tests/providers/airbyte/__init__.py",),
             {
                 "affected-providers-list-as-string": "airbyte",
-                "all-python-versions": "['3.8']",
-                "all-python-versions-list-as-string": "3.8",
-                "python-versions": "['3.8']",
-                "python-versions-list-as-string": "3.8",
+                "all-python-versions": "['3.9']",
+                "all-python-versions-list-as-string": "3.9",
+                "python-versions": "['3.9']",
+                "python-versions-list-as-string": "3.9",
                 "ci-image-build": "true",
                 "prod-image-build": "false",
                 "needs-helm-tests": "false",
@@ -653,10 +652,10 @@ def assert_outputs_are_printed(expected_outputs: dict[str, str], stderr: str):
                 "affected-providers-list-as-string": "amazon apache.hive cncf.kubernetes "
                 "common.compat common.sql exasol ftp google http imap microsoft.azure "
                 "mongo mysql openlineage postgres salesforce ssh teradata",
-                "all-python-versions": "['3.8']",
-                "all-python-versions-list-as-string": "3.8",
-                "python-versions": "['3.8']",
-                "python-versions-list-as-string": "3.8",
+                "all-python-versions": "['3.9']",
+                "all-python-versions-list-as-string": "3.9",
+                "python-versions": "['3.9']",
+                "python-versions-list-as-string": "3.9",
                 "ci-image-build": "true",
                 "prod-image-build": "false",
                 "needs-helm-tests": "false",
@@ -683,10 +682,10 @@ def assert_outputs_are_printed(expected_outputs: dict[str, str], stderr: str):
             ),
             {
                 "affected-providers-list-as-string": "common.compat common.io openlineage",
-                "all-python-versions": "['3.8']",
-                "all-python-versions-list-as-string": "3.8",
-                "python-versions": "['3.8']",
-                "python-versions-list-as-string": "3.8",
+                "all-python-versions": "['3.9']",
+                "all-python-versions-list-as-string": "3.9",
+                "python-versions": "['3.9']",
+                "python-versions-list-as-string": "3.9",
                 "ci-image-build": "true",
                 "prod-image-build": "false",
                 "needs-helm-tests": "false",
@@ -707,10 +706,10 @@ def assert_outputs_are_printed(expected_outputs: dict[str, str], stderr: str):
             ("airflow/operators/bash.py",),
             {
                 "affected-providers-list-as-string": None,
-                "all-python-versions": "['3.8']",
-                "all-python-versions-list-as-string": "3.8",
-                "python-versions": "['3.8']",
-                "python-versions-list-as-string": "3.8",
+                "all-python-versions": "['3.9']",
+                "all-python-versions-list-as-string": "3.9",
+                "python-versions": "['3.9']",
+                "python-versions-list-as-string": "3.9",
                 "ci-image-build": "true",
                 "prod-image-build": "false",
                 "needs-helm-tests": "false",
@@ -731,10 +730,10 @@ def assert_outputs_are_printed(expected_outputs: dict[str, str], stderr: str):
             ("tests/operators/bash.py",),
             {
                 "affected-providers-list-as-string": None,
-                "all-python-versions": "['3.8']",
-                "all-python-versions-list-as-string": "3.8",
-                "python-versions": "['3.8']",
-                "python-versions-list-as-string": "3.8",
+                "all-python-versions": "['3.9']",
+                "all-python-versions-list-as-string": "3.9",
+                "python-versions": "['3.9']",
+                "python-versions-list-as-string": "3.9",
                 "ci-image-build": "true",
                 "prod-image-build": "false",
                 "needs-helm-tests": "false",
@@ -756,10 +755,10 @@ def assert_outputs_are_printed(expected_outputs: dict[str, str], stderr: str):
                 ("tests/utils/test_cli_util.py",),
                 {
                     "affected-providers-list-as-string": ALL_PROVIDERS_AFFECTED,
-                    "all-python-versions": "['3.8']",
-                    "all-python-versions-list-as-string": "3.8",
-                    "python-versions": "['3.8']",
-                    "python-versions-list-as-string": "3.8",
+                    "all-python-versions": "['3.9']",
+                    "all-python-versions-list-as-string": "3.9",
+                    "python-versions": "['3.9']",
+                    "python-versions-list-as-string": "3.9",
                     "ci-image-build": "true",
                     "prod-image-build": "true",
                     "needs-helm-tests": "true",
@@ -855,9 +854,7 @@ def test_excluded_providers():
     )
     assert_outputs_are_printed(
         {
-            "excluded-providers-as-string": json.dumps(
-                {"3.8": ["cloudant"], "3.9": ["cloudant"], "3.12": ["apache.beam"]}
-            ),
+            "excluded-providers-as-string": json.dumps({"3.9": ["cloudant"], "3.12": ["apache.beam"]}),
         },
         str(stderr),
     )
@@ -925,15 +922,15 @@ def test_full_test_needed_when_scripts_changes(files: tuple[str, ...], expected_
                 {
                     "affected-providers-list-as-string": ALL_PROVIDERS_AFFECTED,
                     "all-versions": "true",
-                    "all-python-versions": "['3.8', '3.9', '3.10', '3.11', '3.12']",
-                    "all-python-versions-list-as-string": "3.8 3.9 3.10 3.11 3.12",
+                    "all-python-versions": "['3.9', '3.10', '3.11', '3.12']",
+                    "all-python-versions-list-as-string": "3.9 3.10 3.11 3.12",
                     "mysql-versions": "['8.0', '8.4']",
                     "postgres-versions": "['12', '13', '14', '15', '16', '17']",
-                    "python-versions": "['3.8', '3.9', '3.10', '3.11', '3.12']",
-                    "python-versions-list-as-string": "3.8 3.9 3.10 3.11 3.12",
+                    "python-versions": "['3.9', '3.10', '3.11', '3.12']",
+                    "python-versions-list-as-string": "3.9 3.10 3.11 3.12",
                     "kubernetes-versions": "['v1.28.13', 'v1.29.8', 'v1.30.4', 'v1.31.0']",
                     "kubernetes-versions-list-as-string": "v1.28.13 v1.29.8 v1.30.4 v1.31.0",
-                    "kubernetes-combos-list-as-string": "3.8-v1.28.13 3.9-v1.29.8 3.10-v1.30.4 3.11-v1.31.0 3.12-v1.28.13",
+                    "kubernetes-combos-list-as-string": "3.9-v1.28.13 3.10-v1.29.8 3.11-v1.30.4 3.12-v1.31.0",
                     "ci-image-build": "true",
                     "prod-image-build": "true",
                     "run-tests": "true",
@@ -958,16 +955,16 @@ def test_full_test_needed_when_scripts_changes(files: tuple[str, ...], expected_
                 "main",
                 {
                     "affected-providers-list-as-string": ALL_PROVIDERS_AFFECTED,
-                    "all-python-versions": "['3.8']",
-                    "all-python-versions-list-as-string": "3.8",
+                    "all-python-versions": "['3.9']",
+                    "all-python-versions-list-as-string": "3.9",
                     "all-versions": "false",
                     "mysql-versions": "['8.0']",
                     "postgres-versions": "['12']",
-                    "python-versions": "['3.8']",
-                    "python-versions-list-as-string": "3.8",
+                    "python-versions": "['3.9']",
+                    "python-versions-list-as-string": "3.9",
                     "kubernetes-versions": "['v1.28.13']",
                     "kubernetes-versions-list-as-string": "v1.28.13",
-                    "kubernetes-combos-list-as-string": "3.8-v1.28.13",
+                    "kubernetes-combos-list-as-string": "3.9-v1.28.13",
                     "ci-image-build": "true",
                     "prod-image-build": "true",
                     "run-tests": "true",
@@ -992,16 +989,16 @@ def test_full_test_needed_when_scripts_changes(files: tuple[str, ...], expected_
                 "main",
                 {
                     "affected-providers-list-as-string": ALL_PROVIDERS_AFFECTED,
-                    "all-python-versions": "['3.8']",
-                    "all-python-versions-list-as-string": "3.8",
+                    "all-python-versions": "['3.9']",
+                    "all-python-versions-list-as-string": "3.9",
                     "all-versions": "false",
                     "mysql-versions": "['8.0']",
                     "postgres-versions": "['12']",
-                    "python-versions": "['3.8']",
-                    "python-versions-list-as-string": "3.8",
+                    "python-versions": "['3.9']",
+                    "python-versions-list-as-string": "3.9",
                     "kubernetes-versions": "['v1.28.13']",
                     "kubernetes-versions-list-as-string": "v1.28.13",
-                    "kubernetes-combos-list-as-string": "3.8-v1.28.13",
+                    "kubernetes-combos-list-as-string": "3.9-v1.28.13",
                     "ci-image-build": "true",
                     "prod-image-build": "true",
                     "run-tests": "true",
@@ -1064,14 +1061,14 @@ def test_full_test_needed_when_scripts_changes(files: tuple[str, ...], expected_
                 "main",
                 {
                     "affected-providers-list-as-string": ALL_PROVIDERS_AFFECTED,
-                    "all-python-versions": "['3.8']",
-                    "all-python-versions-list-as-string": "3.8",
+                    "all-python-versions": "['3.9']",
+                    "all-python-versions-list-as-string": "3.9",
                     "all-versions": "false",
-                    "python-versions": "['3.8']",
-                    "python-versions-list-as-string": "3.8",
+                    "python-versions": "['3.9']",
+                    "python-versions-list-as-string": "3.9",
                     "kubernetes-versions": "['v1.28.13']",
                     "kubernetes-versions-list-as-string": "v1.28.13",
-                    "kubernetes-combos-list-as-string": "3.8-v1.28.13",
+                    "kubernetes-combos-list-as-string": "3.9-v1.28.13",
                     "ci-image-build": "true",
                     "prod-image-build": "true",
                     "run-tests": "true",
@@ -1096,14 +1093,14 @@ def test_full_test_needed_when_scripts_changes(files: tuple[str, ...], expected_
                 "main",
                 {
                     "affected-providers-list-as-string": ALL_PROVIDERS_AFFECTED,
-                    "all-python-versions": "['3.8']",
-                    "all-python-versions-list-as-string": "3.8",
+                    "all-python-versions": "['3.9']",
+                    "all-python-versions-list-as-string": "3.9",
                     "all-versions": "false",
-                    "python-versions": "['3.8']",
-                    "python-versions-list-as-string": "3.8",
+                    "python-versions": "['3.9']",
+                    "python-versions-list-as-string": "3.9",
                     "kubernetes-versions": "['v1.28.13']",
                     "kubernetes-versions-list-as-string": "v1.28.13",
-                    "kubernetes-combos-list-as-string": "3.8-v1.28.13",
+                    "kubernetes-combos-list-as-string": "3.9-v1.28.13",
                     "ci-image-build": "true",
                     "prod-image-build": "true",
                     "run-tests": "true",
@@ -1132,10 +1129,10 @@ def test_full_test_needed_when_scripts_changes(files: tuple[str, ...], expected_
                 "v2-7-stable",
                 {
                     "affected-providers-list-as-string": ALL_PROVIDERS_AFFECTED,
-                    "all-python-versions": "['3.8']",
-                    "all-python-versions-list-as-string": "3.8",
-                    "python-versions": "['3.8']",
-                    "python-versions-list-as-string": "3.8",
+                    "all-python-versions": "['3.9']",
+                    "all-python-versions-list-as-string": "3.9",
+                    "python-versions": "['3.9']",
+                    "python-versions-list-as-string": "3.9",
                     "all-versions": "false",
                     "ci-image-build": "true",
                     "prod-image-build": "true",
@@ -1184,8 +1181,8 @@ def test_expected_output_full_tests_needed(
             ("INTHEWILD.md",),
             {
                 "affected-providers-list-as-string": None,
-                "all-python-versions": "['3.8']",
-                "all-python-versions-list-as-string": "3.8",
+                "all-python-versions": "['3.9']",
+                "all-python-versions-list-as-string": "3.9",
                 "ci-image-build": "false",
                 "needs-helm-tests": "false",
                 "run-tests": "false",
@@ -1209,8 +1206,8 @@ def test_expected_output_full_tests_needed(
                 "affected-providers-list-as-string": "amazon apache.beam apache.cassandra cncf.kubernetes "
                 "common.compat common.sql facebook google hashicorp microsoft.azure microsoft.mssql "
                 "mysql openlineage oracle postgres presto salesforce samba sftp ssh trino",
-                "all-python-versions": "['3.8']",
-                "all-python-versions-list-as-string": "3.8",
+                "all-python-versions": "['3.9']",
+                "all-python-versions-list-as-string": "3.9",
                 "needs-helm-tests": "false",
                 "ci-image-build": "true",
                 "prod-image-build": "true",
@@ -1239,8 +1236,8 @@ def test_expected_output_full_tests_needed(
                 "cncf.kubernetes common.compat common.sql facebook google "
                 "hashicorp microsoft.azure microsoft.mssql mysql openlineage oracle postgres "
                 "presto salesforce samba sftp ssh trino",
-                "all-python-versions": "['3.8']",
-                "all-python-versions-list-as-string": "3.8",
+                "all-python-versions": "['3.9']",
+                "all-python-versions-list-as-string": "3.9",
                 "ci-image-build": "true",
                 "prod-image-build": "true",
                 "needs-helm-tests": "false",
@@ -1264,8 +1261,8 @@ def test_expected_output_full_tests_needed(
             ),
             {
                 "affected-providers-list-as-string": ALL_PROVIDERS_AFFECTED,
-                "all-python-versions": "['3.8']",
-                "all-python-versions-list-as-string": "3.8",
+                "all-python-versions": "['3.9']",
+                "all-python-versions-list-as-string": "3.9",
                 "ci-image-build": "true",
                 "prod-image-build": "false",
                 "needs-helm-tests": "false",
@@ -1307,8 +1304,8 @@ def test_expected_output_pull_request_v2_7(
             ("INTHEWILD.md",),
             {
                 "affected-providers-list-as-string": None,
-                "all-python-versions": "['3.8']",
-                "all-python-versions-list-as-string": "3.8",
+                "all-python-versions": "['3.9']",
+                "all-python-versions-list-as-string": "3.9",
                 "ci-image-build": "false",
                 "needs-helm-tests": "false",
                 "run-tests": "false",
@@ -1328,8 +1325,8 @@ def test_expected_output_pull_request_v2_7(
             ("tests/system/any_file.py",),
             {
                 "affected-providers-list-as-string": None,
-                "all-python-versions": "['3.8']",
-                "all-python-versions-list-as-string": "3.8",
+                "all-python-versions": "['3.9']",
+                "all-python-versions-list-as-string": "3.9",
                 "ci-image-build": "true",
                 "prod-image-build": "false",
                 "needs-helm-tests": "false",
@@ -1357,8 +1354,8 @@ def test_expected_output_pull_request_v2_7(
                 "cncf.kubernetes common.compat common.sql "
                 "facebook google hashicorp microsoft.azure microsoft.mssql mysql "
                 "openlineage oracle postgres presto salesforce samba sftp ssh trino",
-                "all-python-versions": "['3.8']",
-                "all-python-versions-list-as-string": "3.8",
+                "all-python-versions": "['3.9']",
+                "all-python-versions-list-as-string": "3.9",
                 "ci-image-build": "true",
                 "prod-image-build": "true",
                 "needs-helm-tests": "true",
@@ -1391,8 +1388,8 @@ def test_expected_output_pull_request_v2_7(
             ),
             {
                 "affected-providers-list-as-string": "common.compat fab",
-                "all-python-versions": "['3.8']",
-                "all-python-versions-list-as-string": "3.8",
+                "all-python-versions": "['3.9']",
+                "all-python-versions-list-as-string": "3.9",
                 "ci-image-build": "true",
                 "prod-image-build": "false",
                 "needs-helm-tests": "false",
@@ -1413,8 +1410,8 @@ def test_expected_output_pull_request_v2_7(
         pytest.param(
             ("airflow/models/test.py",),
             {
-                "all-python-versions": "['3.8']",
-                "all-python-versions-list-as-string": "3.8",
+                "all-python-versions": "['3.9']",
+                "all-python-versions-list-as-string": "3.9",
                 "ci-image-build": "true",
                 "prod-image-build": "false",
                 "needs-helm-tests": "false",
@@ -1435,8 +1432,8 @@ def test_expected_output_pull_request_v2_7(
         pytest.param(
             ("airflow/file.py",),
             {
-                "all-python-versions": "['3.8']",
-                "all-python-versions-list-as-string": "3.8",
+                "all-python-versions": "['3.9']",
+                "all-python-versions-list-as-string": "3.9",
                 "ci-image-build": "true",
                 "prod-image-build": "false",
                 "needs-helm-tests": "false",
@@ -1480,8 +1477,8 @@ def test_expected_output_pull_request_target(
             "main",
             {
                 "affected-providers-list-as-string": ALL_PROVIDERS_AFFECTED,
-                "all-python-versions": "['3.8', '3.9', '3.10', '3.11', '3.12']",
-                "all-python-versions-list-as-string": "3.8 3.9 3.10 3.11 3.12",
+                "all-python-versions": "['3.9', '3.10', '3.11', '3.12']",
+                "all-python-versions-list-as-string": "3.9 3.10 3.11 3.12",
                 "ci-image-build": "true",
                 "prod-image-build": "true",
                 "needs-helm-tests": "true",
@@ -1502,8 +1499,8 @@ def test_expected_output_pull_request_target(
             "v2-3-stable",
             {
                 "affected-providers-list-as-string": ALL_PROVIDERS_AFFECTED,
-                "all-python-versions": "['3.8', '3.9', '3.10', '3.11', '3.12']",
-                "all-python-versions-list-as-string": "3.8 3.9 3.10 3.11 3.12",
+                "all-python-versions": "['3.9', '3.10', '3.11', '3.12']",
+                "all-python-versions-list-as-string": "3.9 3.10 3.11 3.12",
                 "ci-image-build": "true",
                 "prod-image-build": "true",
                 "needs-helm-tests": "false",
@@ -1526,8 +1523,8 @@ def test_expected_output_pull_request_target(
             "main",
             {
                 "affected-providers-list-as-string": ALL_PROVIDERS_AFFECTED,
-                "all-python-versions": "['3.8', '3.9', '3.10', '3.11', '3.12']",
-                "all-python-versions-list-as-string": "3.8 3.9 3.10 3.11 3.12",
+                "all-python-versions": "['3.9', '3.10', '3.11', '3.12']",
+                "all-python-versions-list-as-string": "3.9 3.10 3.11 3.12",
                 "ci-image-build": "true",
                 "prod-image-build": "true",
                 "needs-helm-tests": "true",
@@ -1580,8 +1577,8 @@ def test_no_commit_provided_trigger_full_build_for_any_event_type(github_event):
     )
     assert_outputs_are_printed(
         {
-            "all-python-versions": "['3.8', '3.9', '3.10', '3.11', '3.12']",
-            "all-python-versions-list-as-string": "3.8 3.9 3.10 3.11 3.12",
+            "all-python-versions": "['3.9', '3.10', '3.11', '3.12']",
+            "all-python-versions-list-as-string": "3.9 3.10 3.11 3.12",
             "ci-image-build": "true",
             "prod-image-build": "true",
             "needs-helm-tests": "true",
